#309: move Peer and PeersRepository to domain peer package
Peer and PeersRepository belong to the domain layer — they represent
discovered network peers, not presentation constructs. Moving them out
of presentation.peer to peer removes the cross-layer citation in
AutoSendDispatcher and AppContainer. All call sites updated; the
presentation.peer directory is removed.
