+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/transfer/PeerTransferComponent.kt
@@ -6,7 +6,7 @@ import com.arkivanov.decompose.value.Value
 import com.arkivanov.decompose.value.update
 import com.arkivanov.essenty.lifecycle.LifecycleRegistry
 import com.arkivanov.essenty.lifecycle.destroy
-import com.tubetoast.tether.presentation.peer.Peer
+import com.tubetoast.tether.peer.Peer
 import com.tubetoast.tether.transfer.FileSource
 import com.tubetoast.tether.transfer.PeerIdentity
 import com.tubetoast.tether.transfer.PeerTransferEngine
