+++ composeApp/src/commonTest/kotlin/com/tubetoast/tether/peer/FakePeersRepository.kt
@@ -1,4 +1,4 @@
-package com.tubetoast.tether.presentation.peer
+package com.tubetoast.tether.peer
 
 import com.tubetoast.tether.discovery.FakeDeviceDiscovery
 import kotlinx.coroutines.CoroutineScope
